[feat]메뉴+수익관리+순수익 1차 머지

+++ Projects/Soonsu/Model/DailySales.swift
@@ -10,6 +10,7 @@ import Foundation
 /// 하루 치 매출/원가 합산 정보
 struct DailySales: Codable, Hashable {
     let revenue: Int
-    let materialCost: Int
+//    let materialCost: Int
+    let materialCost: Double
     let items: [SoldItem]
 }
